cloud-functions/text-recognizer/src/requirements.txt
google-cloud-storage
google-cloud-vision
google-cloud-pubsub
google-cloud-firestore